refactor: move endpoint specific logic to an endpoints dir
