+++ api/server/routes/endpoints/gptPlugins/buildOptions.js
@@ -1,23 +1,25 @@
 const { parseConvo } = require('librechat-data-provider');
-const { handleError } = require('../../utils');
+const { handleError } = require('../../../utils');
 
 const buildOptions = (req, res) => {
   const { endpoint, text, parentMessageId, conversationId } = req.body;
   if (text.length === 0) {
     return handleError(res, { text: 'Prompt empty or too short' });
   }
-  const isOpenAI = endpoint === 'openAI' || endpoint === 'azureOpenAI';
-  if (!isOpenAI) {
+  if (endpoint !== 'gptPlugins') {
     return handleError(res, { text: 'Illegal request' });
   }
 
+  const tools = req.body?.tools.map((tool) => tool.pluginKey) ?? [];
   // build endpoint option
   const parsedBody = parseConvo(endpoint, req.body);
-  const { chatGptLabel, promptPrefix, ...rest } = parsedBody;
+  const { chatGptLabel, promptPrefix, agentOptions, ...rest } = parsedBody;
   const endpointOption = {
+    tools,
     endpoint,
     chatGptLabel,
     promptPrefix,
+    agentOptions,
     modelOptions: {
       ...rest,
     },
