+++ api/server/routes/endpoints/openAI/buildOptions.js
@@ -1,25 +1,23 @@
 const { parseConvo } = require('librechat-data-provider');
-const { handleError } = require('../../utils');
+const { handleError } = require('../../../utils');
 
 const buildOptions = (req, res) => {
   const { endpoint, text, parentMessageId, conversationId } = req.body;
   if (text.length === 0) {
     return handleError(res, { text: 'Prompt empty or too short' });
   }
-  if (endpoint !== 'gptPlugins') {
+  const isOpenAI = endpoint === 'openAI' || endpoint === 'azureOpenAI';
+  if (!isOpenAI) {
     return handleError(res, { text: 'Illegal request' });
   }
 
-  const tools = req.body?.tools.map((tool) => tool.pluginKey) ?? [];
   // build endpoint option
   const parsedBody = parseConvo(endpoint, req.body);
-  const { chatGptLabel, promptPrefix, agentOptions, ...rest } = parsedBody;
+  const { chatGptLabel, promptPrefix, ...rest } = parsedBody;
   const endpointOption = {
-    tools,
     endpoint,
     chatGptLabel,
     promptPrefix,
-    agentOptions,
     modelOptions: {
       ...rest,
     },
